1. 修复加载配置文件时会因为旧的配置文件没有新的配置项而报KeyError的问题
